requirements.txt
pdfplumber
python-docx
requests
python-dotenv
google-cloud-texttospeech
google-generativeai
gTTS
